React joins/leaves (#4)
* Begin work on reaction joins/leaves

* Fix leave response message

* Only let users leave groups they are members of

+++ src/actions/leave.ts
@@ -2,11 +2,20 @@ import { getRepository } from 'typeorm';
 import { Event } from '../entity/Event';
 import { Player } from '../entity/Player';
 
+export enum LEAVE_RESULT {
+  LEFT_EVENT,
+  NOT_IN_EVENT,
+  TRANSFERRED,
+  EVENT_REMOVED,
+}
+
 export const leaveEvent = async function leaveEvent({
   discordId,
+  guildId,
 }: {
   discordId: string;
-}): Promise<string> {
+  guildId: string;
+}): Promise<{ result: LEAVE_RESULT; msg: string }> {
   const playerRepository = getRepository(Player);
   const eventRepository = getRepository(Event);
   const player = await playerRepository.findOne({
@@ -18,15 +27,21 @@ export const leaveEvent = async function leaveEvent({
     throw new Error('Player data missing when calling leave event');
   }
 
-  if (!player.joinedEvent) {
-    return 'You are not currently in an event.';
+  if (!player.joinedEvent || player.joinedEvent.guildId !== guildId) {
+    return {
+      result: LEAVE_RESULT.NOT_IN_EVENT,
+      msg: 'You are not currently in an event.',
+    };
   }
 
   if (player.joinedEvent.owner.id === player.id) {
     // Owner is leaving, switch ownership to next player
     if (player.joinedEvent.players.length === 1) {
       await eventRepository.delete(player.joinedEvent.id);
-      return `**${player.joinedEvent.name}** has been removed`;
+      return {
+        result: LEAVE_RESULT.EVENT_REMOVED,
+        msg: `**${player.joinedEvent.name}** has been removed`,
+      };
     } else {
       const newOwner = player.joinedEvent.players.find(
         (p) => p.id !== player.id
@@ -36,23 +51,30 @@ export const leaveEvent = async function leaveEvent({
       );
 
       if (!newOwner || newPlayers.length === 0) {
-        await eventRepository.delete(player.joinedEvent);
-        return `**${player.joinedEvent.name}** has been removed`;
+        await eventRepository.delete(player.joinedEvent.id);
+        return {
+          result: LEAVE_RESULT.EVENT_REMOVED,
+          msg: `**${player.joinedEvent.name}** has been removed`,
+        };
       }
       player.joinedEvent.players = newPlayers;
       player.joinedEvent.owner = newOwner;
 
       let msg = `**${player.name}** has left event: **${player.joinedEvent.name}**.\n`;
-      msg += `Ownership transferred to: **${newOwner}**`;
-      return msg;
+      msg += `Ownership transferred to: **${newOwner.name}**`;
+      return { result: LEAVE_RESULT.TRANSFERRED, msg };
     }
   }
+
   const newPlayers = player.joinedEvent.players.filter(
     (p) => player.id !== p.id
   );
   player.joinedEvent.players = newPlayers;
-  eventRepository.save(newPlayers);
-  return `**${player.name}** has left group: **${player.joinedEvent.name}**`;
+  await eventRepository.save(player.joinedEvent);
+  return {
+    result: LEAVE_RESULT.LEFT_EVENT,
+    msg: `**${player.name}** has left group: **${player.joinedEvent.name}**`,
+  };
 };
 
 export default leaveEvent;
